requirements.txt
braceexpand==0.1.7
certifi==2025.6.15
charset-normalizer==3.4.2
contourpy==1.3.2
cycler==0.12.1
facenet-pytorch==2.6.0
filelock==3.18.0
fonttools==4.58.4
fsspec==2025.5.1
hf-xet==1.1.5
huggingface-hub==0.33.1
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
opencv-python==4.11.0.86
packaging==25.0
pandas==2.3.0
pillow==10.2.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
six==1.17.0
sympy==1.14.0
threadpoolctl==3.6.0
timm==1.0.16
torch==2.2.2
torchvision==0.17.2
tqdm==4.67.1
triton==2.2.0
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
webdataset==1.0.2
